CLean up docs + code, add shared.ts
