+++ src/shared.ts
@@ -12,8 +12,17 @@ export const getIngestURL = (t: EndpointType) => {
   if (!ingestEndpoint) {
     return '';
   }
+
   const url = new URL(ingestEndpoint);
   // attach type query param based on passed EndpointType
   url.searchParams.set('type', t.toString());
   return url.toString();
+};
+
+export const debounce = (fn: Function, ms = 300) => {
+  let timeoutId: ReturnType<typeof setTimeout>;
+  return async function (this: any, ...args: any[]) {
+    clearTimeout(timeoutId);
+    timeoutId = setTimeout(async () => await fn.apply(this, args), ms);
+  };
 };
